console,oauth: Rename app to init for redux primitives related to initializing the apps
