+++ pkg/webui/console/store/actions/init.js
