+++ pkg/webui/console/store/middleware/index.js
@@ -14,7 +14,7 @@
 
 import client from './client'
 import user from './user'
-import app from './app'
+import init from './init'
 import applications from './applications'
 import application from './application'
 import devices from './devices'
@@ -23,7 +23,7 @@ import gateways from './gateways'
 export default [
   client,
   ...user,
-  app,
+  init,
   ...applications,
   ...application,
   ...devices,
